requirements.txt
google-adk
google-cloud-bigquey
ipykernel
pandas
db-dtypes